Add dtype to LDAModel. Partially fix (#1656)
* Add dtype to LdaModel, assert about it everywhere

* Implement loading of old models without dtype

* Use assert_allclose instead of == in LdaModel tests. Use np.issubdtype when checking if something is float.

* Fix AuthorTopicModel

* Fix matutils.dirichlet_expectation

 * replace assert with docstring comment
 * add test to check that it really saves dtype for different inputs

* Change default to np.float32 and cleanup

* Fix wrong logging message

* Remove out-of-place comment

* Cleanup PEP8

* Add dtype to sklearn LdaTransformer

* Set precision explicitly in lda model converters

* Add dtype to LdaMulticore

* Set dtype to float64 explicitly to retain backward compatibility in models using LdaModel

* Cleanup asserts and fix another place calculating in float64

* Fix import

* Fix remarks by piskvorky

* Add backward compatibility tests

* Add missing .npy files

* Fix dirichlet_expectation not working with np.float16

* Fix path to saved model

+++ gensim/matutils.py
@@ -602,13 +602,12 @@ def jaccard_distance(set1, set2):
 def dirichlet_expectation(alpha):
     """
     For a vector `theta~Dir(alpha)`, compute `E[log(theta)]`.
-
     """
     if len(alpha.shape) == 1:
         result = psi(alpha) - psi(np.sum(alpha))
     else:
         result = psi(alpha) - psi(np.sum(alpha, 1))[:, np.newaxis]
-    return result.astype(alpha.dtype)  # keep the same precision as input
+    return result.astype(alpha.dtype, copy=False)  # keep the same precision as input
 
 
 def qr_destroy(la):
